added auth authorization to edit and delete

+++ rest-api/controllers/feed.js
@@ -132,6 +132,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw Error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        // Not authorized
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -159,6 +165,12 @@ exports.deletePost = (req, res, next) => {
         const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw Error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        // Not authorized
+        error.statusCode = 403;
+        throw error;
       }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
